Correction des bugs. Passage des types de notes en doubles car des notes pourraient être des valeurs décimals. Cela corrige aussi en parti les calcules de moyennes.

+++ src/fr/group3il/tp6/Student.java
@@ -2,9 +2,9 @@ package fr.group3il.tp6;
 
 public class Student {
     public String name;
-    public int[] grades;
+    public double[] grades;
 
-    public Student(String name, int[] grades) {
+    public Student(String name, double[] grades) {
         this.name = name;
         this.grades = grades;
     }
